[LayoutNG] Small legend large border and large legend small border
Added two web tests:
1. Small legend large border: At this time, the legend is treated as
monolithic and the borders are not breaking as well. As a result, the
border goes past the fragmentation line in the first and last columns.
2. Large legend small border: Similarly, the legend goes past the
fragmentation line in the first column. The fieldset breaks when it
reaches the fragmentation line in the second column.

Deleted the update of the border block start value as it's no longer
useful to calculate new offsets in the fieldset.

Bug: 829028
Change-Id: I1746c9da67372f4a42ab7ab8c4184460814f486d